requirements.txt
pandas
numpy
matplotlib
ta
yfinance
schedule
requests
pyyaml